requirements.txt
Django==4.2.20
djangorestframework
djangorestframework-simplejwt
Pillow
